Move WorkflowWorker from workers.py to workflow.py service module
